// frontend/src/api/generated/models/createTradeCommand.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Fusion Trade Blotter API
 * OpenAPI spec version: 1.0.0
 */
import type { TradeSide } from './tradeSide';

export interface CreateTradeCommand {
  /**
     * @minLength 1
     * @maxLength 12
     */
  symbol: string;
  side: TradeSide;
  /**
     * @maximum 1000000000
     * @exclusiveMinimum 0
     */
  quantity: number;
  /**
     * @maximum 1000000000
     * @exclusiveMinimum 0
     */
  price: number;
  /**
     * @minLength 1
     * @maxLength 40
     */
  book: string;
  /**
     * @minLength 1
     * @maxLength 80
     */
  counterparty: string;
  /** @pattern ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d:[0-5]\d(?:\.\d+)?(?:Z))$ */
  tradeTimestamp: string;
}